feat: :sparkles: Exercise 10 Solution
In this point  Declare an empty array "myArray", Set "addNum" to 1 as a starting value, Loop through the input process until the user enters a 0 as input.  Use parseInt() to convert the user input to an integer and push it to "myArray". delete the last number that push into "myArray" already it eas "0" with ".pop" m,ethod. Calculate the sum of the numbers in myArray using the reduce() method and store it in the add variable.  Calculate the average of the numbers in myArray by dividing add by the length of myArray, and store it in the "prom" variable.  Find the maximum value of the numbers in myArray using the spread operator and the Math.max() method, and store it in the "max" variable.  Find the minimum value of the numbers in myArray using the spread operator and the Math.min() method, and store it in the "min" variable,  Log the results to the console using console.log() method.